pascoli/backend/requirements.txt
psycopg2
Flask
Flask-SQLAlchemy
PyJWT
python-decouple
pydantic[email]